VUELOS/requirements.txt
django
django-extensions
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests